requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.2
billiard==3.6.3.0
braintree==3.59.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
django-celery-beat==1.1.1
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
stripe==2.43.0
suds-jurko==0.6
urllib3==1.25.8
vine==1.3.0
xmltodict==0.12.0
zipp==3.1.0
Django==3.0.6
djangorestframework==3.11.0